requirements.txt
-i https://pypi.org/simple
absl-py==0.10.0
aiodns==1.1.1 ; python_version >= '3.5.2'
aiohttp==3.6.3 ; python_version >= '3.5.2'
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.2.0
cachetools==4.1.1
ccxt==1.36.82
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cloudpickle==1.6.0
cryptography==3.1.1
cycler==0.10.0
decorator==4.4.2
flatlib==0.2.1
future==0.18.2
gast==0.3.3
google-auth-oauthlib==0.4.1
google-auth==1.22.1
google-pasta==0.2.0
grpcio==1.33.1
h5py==2.10.0
hyperopt==0.2.5
idna==2.8
importlib-metadata==2.0.0 ; python_version < '3.8'
inflection==0.5.1
joblib==0.17.0
keras-preprocessing==1.1.2
kiwisolver==1.2.0
lxml==4.6.1
markdown==3.3.2
matplotlib==3.1.0
more-itertools==8.5.0
multidict==4.7.6
networkx==2.5
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas-datareader==0.9.0
pandas==0.25.0
protobuf==3.13.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycares==3.1.1
pycparser==2.20
pyparsing==2.4.7
pyswisseph==2.0.0.post2
python-dateutil==2.8.1
python-telegram-bot==11.1.0
pytz==2020.1
quandl==3.5.3
requests-oauthlib==1.3.0
requests==2.22.0
rsa==4.6 ; python_version >= '3.5'
scikit-learn==0.21.3
scipy==1.5.3
six==1.15.0
ta-lib==0.4.17
tensorboard-plugin-wit==1.7.0
tensorboard==2.3.0
tensorflow-estimator==2.3.0
tensorflow==2.3.1
termcolor==1.1.0
tqdm==4.50.2
urllib3==1.25.11
werkzeug==1.0.1
wheel==0.35.1 ; python_version >= '3'
wrapt==1.12.1
yarl==1.1.0 ; python_version >= '3.5.2'
zipp==3.3.2
